moving redundant kg generation into Common
